+++ adapters/MoPub/MoPubAdapter/MoPubAdapterConstants.h
@@ -1,3 +1,12 @@
 #define MINIMUM_MOPUB_PRIVACY_ICON_SIZE 10
 #define DEFAULT_MOPUB_PRIVACY_ICON_SIZE 20
 #define MAXIMUM_MOPUB_PRIVACY_ICON_SIZE 30
+
+/// MoPub mediation network adapter version.
+static NSString *const GADMAdapterMoPubVersion = @"5.5.0.0";
+
+/// Constant for adapter error domain.
+static NSString *const kAdapterErrorDomain = @"com.mopub.mobileads.MoPubAdapter";
+
+/// Internal to MoPub
+static NSString *const kAdapterTpValue = @"gmext";