+++ adapters/Nend/NendAdapter/GADMAdapterNendSetting.h
@@ -7,5 +7,5 @@
 
 #ifndef NendAdapter_GADMAdapterNendSetting_h
 #define NendAdapter_GADMAdapterNendSetting_h
-#define GADM_ADAPTER_NEND_VERSION (@"4.0.5.0")
+#define GADM_ADAPTER_NEND_VERSION (@"5.1.0.0")
 #endif