+++ adapters/Vungle/Public/Headers/VungleRouterConsent.h
@@ -1,15 +1,21 @@
+// Copyright 2017 Google Inc.
 //
-//  VungleRouterConsent.h
-//  VungleAdapter
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-//  Created by Vungle on 5/18/18.
-//  Copyright © 2018 Vungle. All rights reserved.
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 #import <Foundation/Foundation.h>
-#import <VungleSDK/VungleSDK.h>
+typedef NS_ENUM(NSInteger, VungleConsentStatus);
 
 @interface VungleRouterConsent : NSObject
-+ (void)updateConsentStatus:(VungleConsentStatus)consentStatus;
++ (BOOL)updateConsentStatus:(VungleConsentStatus)consentStatus;
 + (VungleConsentStatus)getConsentStatus;
 @end