+++ adapters/AppLovin/CHANGELOG.md
@@ -1,5 +1,27 @@
 # AppLovin Adapter for Google Mobile Ads SDK for iOS
 
+## 6.2.0.0
+- Verified compatibility with AppLovin SDK 6.2.0.
+
+## 6.1.4.0
+- Verified compatibility with AppLovin SDK 6.1.4.
+
+## 5.1.2.0
+- Verified compatibility with AppLovin SDK 5.1.2.
+
+## 5.1.1.0
+- Verified compatibility with AppLovin SDK 5.1.1.
+
+## 5.1.0.0
+- Verified compatibility with AppLovin SDK 5.1.0.
+
+## 5.0.2.0
+- Verified compatibility with AppLovin SDK 5.0.2.
+
+## 5.0.1.1
+- Add support for native ads.
+- Set AdMob as mediation provider on the AppLovin SDK.
+
 ## Version 5.0.1.0
 - Verified compatibility with Applovin SDK 5.0.1.