+++ adapters/Facebook/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Facebook Audience Network Mediation Adapter for Google Mobile Ads SDK for iOS
 
+## Version 5.1.1.0
+- Verified compatibility with FAN SDK 5.1.1.
+
+## Version 5.1.0.0
+- Verified compatibility with FAN SDK 5.1.0.
+
+## Version 5.0.1.0
+- Verified compatibility with FAN SDK 5.0.1.
+
+## Version 5.0.0.0
+- Verified compatibility with FAN SDK 5.0.0.
+
+## Version 4.99.3.0
+- Verified compatibility with FAN SDK 4.99.3.
+
+## Version 4.99.2.0
+- Verified compatibility with FAN SDK 4.99.2.
+
+## Version 4.99.1.0
+- Verified compatibility with FAN SDK 4.99.1.
+
+## Version 4.28.1.2
+- Added `adapterDidCompletePlayingRewardBasedVideoAd:` callback to the adapter.
+
 ## Version 4.28.1.1
 - Set mediation service for Facebook adapter.