+++ adapters/InMobi/CHANGELOG.md
@@ -1,5 +1,23 @@
 # InMobi Adapter for Google Mobile Ads SDK for iOS
 
+## Version 7.2.4.0
+- Verified compatibility with InMobi SDK 7.2.4.
+
+## Version 7.2.1.0
+- Verified compatibility with InMobi SDK 7.2.1.
+
+## Version 7.2.0.0
+- Verified compatibility with InMobi SDK 7.2.0.
+
+## Version 7.1.2.0
+- Verified compatibility with InMobi SDK 7.1.2.
+
+## Version 7.1.1.2
+- Added `adapterDidCompletePlayingRewardBasedVideoAd:` callback to the adapter.
+
+## Version 7.1.1.1
+- Added the `GADInMobiConsent` class which provides `updateGDPRConsent` and `getConsent` methods.
+
 ## Version 7.1.1.0
 - Verified compatibility with InMobi SDK 7.1.1.