+++ adapters/MyTarget/CHANGELOG.md
@@ -1,5 +1,31 @@
 # myTarget Adapter for Google Mobile Ads SDK for iOS
 
+## Version 5.0.1.0
+- Verified compatibility with myTarget SDK version 5.0.1.
+
+## Version 4.8.9.0
+- Verified compatibility with myTarget SDK version 4.8.9.
+
+## Version 4.8.5.0
+- Verified compatibility with myTarget SDK version 4.8.5.
+
+## Version 4.8.0.0
+- Verified compatibility with myTarget SDK version 4.8.0.
+
+## Version 4.7.11.0
+- Verified compatibility with myTarget SDK version 4.7.11.
+
+## Version 4.7.10.0
+- Verified compatibility with myTarget SDK version 4.7.10.
+- Added `adapterDidCompletePlayingRewardBasedVideoAd:` callback to the adapter.
+
+## Version 4.7.9.1
+- Added support for Unified native ad.
+
+## Version 4.7.9.0
+- Verified compatibility with myTarget SDK version 4.7.9.
+- Added support for unified native ad.
+
 ## Version 4.7.8.0
 - Verified compatibility with myTarget SDK version 4.7.8.