+++ adapters/Nend/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Nend Adapter for Google Mobile Ads SDK for iOS
 
+## Version 5.1.0.0
+- Verified compatibility with nend SDK 5.1.0.
+
+## Version 5.0.3.0
+- Verified compatibility with nend SDK 5.0.3.
+
+## Version 5.0.2.0
+- Verified compatibility with nend SDK 5.0.2.
+
+## Version 5.0.1.0
+- Verified compatibility with nend SDK 5.0.1.
+
+## Version 5.0.0.0
+- Verified compatibility with nend SDK 5.0.0.
+
+## Version 4.0.6.0
+- Verified compatibility with nend SDK 4.0.6.
+- Remove to set userFeature property.
+
 ## Version 4.0.5.0
 - Verified compatibility with nend SDK 4.0.5.