+++ adapters/Oath/CHANGELOG.md
@@ -1,6 +1,7 @@
-# One by AOL Mediation Adapter for Google Mobile Ads SDK for iOS
+# Oath Mediation Adapter for Google Mobile Ads SDK for iOS
 
 ## Earlier versions
 
 - Supports banner and interstitial ads.
 - Supports rewarded video ads as custom event.
+- 'One by AOL' has been rebranded to 'Oath'.